Fix spurious exception in `on_warmup`.
It seems that a call to `on_warmup` can get dispatched even though
another call to `on_warmup` is not fully finished. So the previous
update of all buffers is not fully completed when we want to set the
updated flag.

This commit just handles the case gracefully, even though the whole
problem should probably get avoided through other means.

Fixes #4.